Updated logout page

+++ src/components/logout.jsx
@@ -1,25 +1,37 @@
 import { useEffect } from "react";
 import axios from "axios";
+
 export const Logout = () => {
   useEffect(() => {
     (async () => {
       try {
-        // eslint-disable-next-line no-unused-vars
-        const { data } = await axios.post(
+        const response = await axios.post(
           `${process.env.REACT_APP_BACKEND_URL}/logout/`,
           {
             refresh_token: localStorage.getItem("refresh_token"),
           },
-          { headers: { "Content-Type": "application/json" } },
-          { withCredentials: true }
+          {
+            headers: { "Content-Type": "application/json" },
+            withCredentials: true,
+          }
         );
-        localStorage.clear();
-        axios.defaults.headers.common["Authorization"] = null;
-        window.location.href = "/login";
-      } catch (e) {
-        console.log("logout not working", e);
+
+        // Check if the response has a 'status' property
+        if (response.status === 200) {
+          // Logout was successful
+          localStorage.clear();
+          axios.defaults.headers.common["Authorization"] = null;
+          window.location.href = "/login";
+        } else {
+          // Handle unsuccessful logout (e.g., display an error message)
+          console.log("Logout not successful", response);
+        }
+      } catch (error) {
+        // Handle any network or other errors
+        console.log("Logout not working", error);
       }
     })();
   }, []);
+
   return <div></div>;
 };
